+++ src/const/blob-types.const.js
@@ -1,0 +1,10 @@
+export const BLOB_TYPES = Object.freeze({
+  assetDescription: 'asset_description',
+  fundOverview: 'fund_overview',
+  fundUpdate: 'fund_update',
+  navUpdate: 'nav_update',
+  fundDocument: 'fund_document',
+  kycSyndicate: 'alpha',
+  kycGeneral: 'kyc_form',
+  bravo: 'bravo'
+})